memory_system/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.4.0
networkx>=3.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
numpy>=1.24.0
tenacity>=8.2.0
tiktoken>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
